Change package names to prepare for generic refactoring
