+++ src/main/java/com/glencoesoftware/bioformats2raw/LimitedQueue.java
@@ -6,7 +6,7 @@
  * missing please request a copy by contacting info@glencoesoftware.com
  */
 
-package com.glencoesoftware.mrxs;
+package com.glencoesoftware.bioformats2raw;
 
 import java.util.concurrent.LinkedBlockingQueue;
